requirements.txt
crewai>=0.14.1
openai>=1.6.1
duckduckgo-search>=4.1.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
